+++ application/src/Auxiliary/AddSlashUrl.php
@@ -2,6 +2,11 @@
 
 namespace src\Auxiliary;
 
+/**
+ * class AddSlashUrl
+ * Classe que adicionar por padrão uma barra no final da URL
+ * para auxiliar na identificação das controllers e métodos
+ */
 class AddSlashUrl
 {
     public function __invoke()