+++ application/src/Controller/Error/GenericErrorController.php
@@ -4,11 +4,16 @@ namespace src\Controller\Error;
 
 use src\Controller\BaseController;
 
+/**
+ * class GenericErrorController
+ * Controller para exibição de página de erro genérica
+ */
 class GenericErrorController extends BaseController
 {
     public function index()
     {
         $dados = [
+            'system_name' => 'Home Sweet Home',
             'titulo' => 'Ocorreu um erro',
             'mensagem' => (!empty($_SESSION['mensagem'])) ? $_SESSION['mensagem'] : null
         ];