+++ application/src/Exception/NotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace src\Exception;
+
+use DomainException;
+
+class NotFoundException extends DomainException
+{
+}